+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 🚀 New Feature
 
 - Add a button to copy all messages from a user history.
+- Add follow status (and follow age if following) in the channel details view.
+- Add the `/followed` command.
 
 ### 🐛 Bug Fix